impl: add common and move baseEntity, converter into it
