+++ house/src/main/kotlin/xyz/bukinator/house/model/House.kt
@@ -10,7 +10,8 @@ import jakarta.persistence.Table
 import org.hibernate.annotations.Comment
 import org.hibernate.annotations.GenericGenerator
 import org.hibernate.annotations.Where
-import xyz.bukinator.house.model.converter.StringToListConverter
+import xyz.bukinator.common.BaseEntity
+import xyz.bukinator.common.converter.StringToListConverter
 import xyz.bukinator.house.model.embeddable.Address
 import xyz.bukinator.house.model.embeddable.Area
 import xyz.bukinator.house.model.embeddable.Floor
