+++ house/src/main/kotlin/xyz/bukinator/house/model/embeddable/Price.kt
@@ -4,7 +4,7 @@ import jakarta.persistence.Column
 import jakarta.persistence.Convert
 import jakarta.persistence.Embeddable
 import org.hibernate.annotations.Comment
-import xyz.bukinator.house.model.converter.StringToListConverter
+import xyz.bukinator.common.converter.StringToListConverter
 
 @Embeddable
 class Price(
